+++ server/migrations_postgres/2_webhooks.down.sql
@@ -1,0 +1,3 @@
+drop table if exists webhooks;
+
+commit;