+++ server/migrations_postgres/4_secrets.down.sql
@@ -1,0 +1,3 @@
+drop table if exists secrets;
+
+commit;