+++ server/migrations_postgres/6_action_invocations.down.sql
@@ -1,0 +1,3 @@
+drop table if exists action_invocations;
+
+commit;